+++ scripts/in_container/verify_providers.py
@@ -256,9 +256,6 @@ KNOWN_DEPRECATED_DIRECT_IMPORTS: set[str] = {
         `airflow.providers.snowflake.transfers.copy_into_snowflake.CopyFromExternalStageToSnowflakeOperator`.
         """
     ),
-    "This module is deprecated. Please use `airflow.providers.atlassian.jira.hooks.jira`.",
-    "This module is deprecated. Please use `airflow.providers.atlassian.jira.operators.jira`.",
-    "This module is deprecated. Please use `airflow.providers.atlassian.jira.sensors.jira`.",
 }